requirements.txt
nose
nrepl-python-client